requirements.txt
pyelftools
pefile
sty
requests
terminaltables
yara-python
ssdeep